+++ eslint.config.mjs
@@ -23,6 +23,6 @@ export default [{
         eqeqeq: "warn",
         "no-throw-literal": "warn",
         semi: "warn",
-        "max-lines": ["warn", { max: 300, skipBlankLines: false, skipComments: false }],
+        "max-lines": ["warn", { max: 300, skipBlankLines: true, skipComments: true }],
     },
 }];